requirements/common.txt
parsel==1.5.*
requests_html==0.10.*
psycopg2-binary==2.8.*
pymongo==3.10.*
